refactor(docs): clean up components.js
