+++ examples/docs/src/components/index.js
@@ -1,0 +1,9 @@
+export { default as theme } from "./theme";
+export { default as mdxProviderComponents } from "./mdxProviderComponents";
+export { default as Code } from "./code";
+export { default as CodeBlock } from "./codeBlock";
+export { default as Container } from "./container";
+export { default as Heading } from "./heading";
+export { default as Notification } from "./notification";
+export { default as Pre } from "./pre";
+export { default as Text } from "./text";
