+++ examples/docs/src/components/theme.js
@@ -1,0 +1,5 @@
+export default {
+  fonts: {
+    mono: '"SF Mono", "Roboto Mono", Menlo, monospace'
+  }
+};
