+++ examples/docs/src/docs-layout.js
@@ -1,35 +1,11 @@
-/* eslint-disable react/jsx-key */
 import React, { Component } from "react";
 import { graphql } from "gatsby";
 
 import MDXRenderer from "gatsby-mdx/mdx-renderer";
 import { MDXProvider } from "@mdx-js/tag";
 
-import Highlight, { defaultProps } from "prism-react-renderer";
 import { ThemeProvider } from "emotion-theming";
-import { Container, components, Pre } from "./components";
-
-const theme = {
-  fonts: {
-    mono: '"SF Mono", "Roboto Mono", Menlo, monospace'
-  }
-};
-
-const CodeBlock = ({ children: exampleCode }) => (
-  <Highlight {...defaultProps} code={exampleCode} language="jsx">
-    {({ className, style, tokens, getLineProps, getTokenProps }) => (
-      <Pre className={className} style={style} p={3}>
-        {tokens.map((line, i) => (
-          <div {...getLineProps({ line, key: i })}>
-            {line.map((token, key) => (
-              <span {...getTokenProps({ token, key })} />
-            ))}
-          </div>
-        ))}
-      </Pre>
-    )}
-  </Highlight>
-);
+import { Container, theme, mdxProviderComponents } from "./components";
 
 export default class MDXRuntimeTest extends Component {
   render() {
@@ -37,12 +13,7 @@ export default class MDXRuntimeTest extends Component {
 
     return (
       <ThemeProvider theme={theme}>
-        <MDXProvider
-          components={{
-            ...components,
-            code: CodeBlock
-          }}
-        >
+        <MDXProvider components={mdxProviderComponents}>
           <Container>
             {children}
             <MDXRenderer tableOfContents={tableOfContents}>
